refactor: centralize crypto utilities
