+++ Cargo.toml
@@ -13,3 +13,4 @@ ed25519-dalek = "2.0"
 rand = "0.8"
 base64 = "0.21"
 tokio = { version = "1", features = ["full"] }
+sonnun_lib = { package = "sonnun", path = "src-tauri" }
